Change deltaTime to be of type double instead of unsigned int.

+++ Source/Engine/Engine.h
@@ -28,7 +28,7 @@ namespace BadEngine {
 			virtual int getRandomNumber(int min, int max) = 0;
 			virtual int getScreenWidth() = 0;
 			virtual int getScreenHeight() = 0;
-			virtual int getCurrentTime() = 0;
+			virtual double getCurrentTime() = 0;
 	};
 };
